+++ sw/airborne/modules/computer_vision/opticflow/opticflow_calculator.h
@@ -61,7 +61,10 @@ struct opticflow_t {
 void opticflow_calc_init(struct opticflow_t *opticflow, uint16_t w, uint16_t h);
 void opticflow_calc_frame(struct opticflow_t *opticflow, struct opticflow_state_t *state, struct image_t *img,
                           struct opticflow_result_t *result);
-
+void calc_fast9_lukas_kanade(struct opticflow_t *opticflow, struct opticflow_state_t *state, struct image_t *img,
+                             struct opticflow_result_t *result);
+void calc_edgeflow(struct opticflow_t *opticflow, struct opticflow_state_t *state, struct image_t *img,
+                   struct opticflow_result_t *result);
 #endif /* OPTICFLOW_CALCULATOR_H */
